Rebased on main and separated EventConsumer integration tests into own test class
